code refactor & minor changes
